\name{abseff}
\alias{abseff}
\title{Hello, World!}
\usage{
hello()
}
\description{
The command abseff was created for the immediate calculating of absolute effects based on relative treatment estimates.

The main intention of this command is to facilitate absolute effects derivation for health researchers and systematic review authors. Editors and peer-reviewers of Cochrane reviews may also find this helpful as they have to revise multiple summary of findings table.
}
\examples{
hello()
}
